Restructure object model
A big change here, but I couldn't work out how to split it into smaller
chunks. After discussions on how best to structure the module, the
approach is now to use object models to register methods with the API.
This should hopefully help to keep the model easy to update in future.